Fix loading custom rules that do not have comment help
Add conditions to check if indexing does not take place on a null valued Value property.

If a PowerShell based custom rule does't have comment based help, the engine throws an error. This occurs because the lack of comment based help results in a null value for the Value property of the dynamically typed description variable.