Add enable parameter to formatting rules
We do not want formatting rules to be run by default whenever scriptanalyzer is invoked. However, the engine doesn't provide a way to disable a given built-in rule without messing with the input parameters or input settings file. Hence we add this parameter to the formatting rules so that they are not run by default whenever scriptanalyer is invoked.