Update UseIdenticalParametersForDSC rule logic
This removes the constraint that a parameter needs to be declared `mandatory` in the `Get/Test/Set` functions if any of them declares the parameter as `mandatory`.

These DSC functions are supposed to take in parameters that declared in the corresponding mof file and only the parameters that are declared as `key` or `required` should be mandatory in all the three functions.