Change resource compilation process
This follows the process that is used in PowerShell Core, except that the directories
in PSSA do not follow the same pattern as PSCore where the assembly name matches the directory
name. When editing resources, you should not use Visual Studio, because it creates the *.Designer.cs file
which is no longer used.
Change the strings for the engine to be named EngineStrings rather than just Strings.
Add the ResGen program to the repo to create the stongly typed string files
Move the resources to a resources directory and remove the Designer.cs files
Add a `-ResGen` parameter to the build script to do on demand resource generation
update appveyor script to build resources