UseApprovedVerbs.md: Improving Documentation (#1294)
* Update UseApprovedVerbs.md

I have come to this documentation a few times in the past. Get-Verb is very helpful, but the additional insight, and examples of unapproved verbs in the Microsoft docs are very helpful and make it easier to decide on a suitable replacement for an unapproved verb. I think it would be helpful to have a link to that documentation here.

* Removed en-us from link

* Clarified text. Broke paragraph into semantic lines.