(MAINT) Fix AvoidUsingDoubleQuotesForConstantString in overview README
As identified[^1] by @BenedekFarkas, the line in the rules overview
readme for `AvoidUsingDoubleQuotesForConstantString` incorrectly lists
the severity as `Warning` and that the rule is configurable when the
severity is `Information`[^2] and there aren't any configurable options.

This change copies the fix @BenedekFarkas kindly submitted to the docs
site to ensure the documentation stays up-to-date.

[^1]: For more information, see:

      - MicrosoftDocs/PowerShell-Docs-Modules#128

[^2]: Source code:

      https://github.com/PowerShell/PSScriptAnalyzer/blob/9a6ce978d870909e688d9da0d101d3fc00b504f0/Rules/AvoidUsingDoubleQuotesForConstantString.cs#L142