Workaround component governance issues
For one, the ADO task that uses `global.json` seems to require that the version use a wildcard.
This is not how it's supposed to be with the `rollForward` option, but this should fix OneBranch using old versions.

For two, we need to opt-out of ADO Dependabot because this is an internal mirror.
The good news is that GitHub Dependabot will be working with CFS again soon.