Move Command Data Files from Engine\Settings -> Engine\CommandDataFiles.
Previously they were mixed in with the Settings Preset files. This wasn't an issue as all settings file were always PSD1 and these command data files are JSON. If we are to support other settings file formats, then we need to separate these.

They are now copied to a CommandDataFiles directory in the built module. Two rules load these and will need to be fixed up to load them from the new directory:

- Rules/AvoidOverwritingBuiltInCmdlets.cs
- Rules/UseCompatibleCmdlets.cs